requirements.txt
streamlit
pandas
reportlab
python-dotenv
openai